requirements.txt
amqp==5.0.6
asgiref==3.4.1
beautifulsoup4==4.3.2
billiard==3.6.4.0
celery==5.2.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-after-response==0.2.2
django-cors-headers==3.7.0
django-extensions==3.1.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-generators==0.5.0
finance==0.2502
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
kombu==5.2.2
lxml==4.6.3
MarkupSafe==2.0.1
multitasking==0.0.9
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.4
pandas-datareader==0.10.0
prompt-toolkit==3.0.23
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
simplejson==3.17.5
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
Unidecode==1.3.2
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
yahoo-finance==1.4.0
yfinance==0.1.64
